Move saveCryptoSetting in Cypher class beside getCryptoSetting
